requirements.txt
langchain
langgraph
langchain-groq
langchain-community
langchain-huggingface

cassio
astrapy

chromadb
tiktoken
wikipedia

pydantic
typing-extensions
python-dotenv

sentence-transformers
